vercel/requirements.txt
# Image Processing
opencv-python-headless==4.8.1.78
pillow==10.1.0

# Barcode Processing
pyzbar==0.1.9

# OpenAI API
openai==1.3.8

# HTTP Requests
requests==2.31.0

# Utilities
numpy==1.24.4
